% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_operations.R
\name{frequency_list}
\alias{frequency_list}
\title{Calculate token frequency list}
\usage{
frequency_list(corpus, pattr = "word", sattr = NULL, sattr_values = NULL)
}
\arguments{
\item{corpus}{corpus created with \code{get_corpus}.}

\item{pattr}{string specifying positional attribute (e.g. word or lemma).}

\item{sattr}{character specifying s-attribute.}

\item{sattr_values}{vector of s-attribute values.}
}
\value{
data.table with tokens, absolute, and relative frequencies.
}
\description{
\code{frequency_list} calculates token (e.g. word or lemma) frequency list
for the specified corpus. If a structural attribute and attribute values are
provided, frequencies are restricted to the subset of the corpus.
}
\examples{
\dontrun{frequency_list, pattr = "lemma", sattr = "text_year", sattr_values = c(2015, 2016)}

}
