% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_operations.R
\name{keywords}
\alias{keywords}
\title{Calculate keywords}
\usage{
keywords(flist_A, flist_B, A_is_subset, min_A = 0, min_B = 0)
}
\arguments{
\item{A_is_subset}{logical. Set to TRUE if counts in freq_list_B contain counts
in freq_list_A. This is the case if the corpus or corpus subset from which
freq_list_A was computed is a subset of the corpus or corpus subset form which
freq_list_B was computed.}

\item{min_A}{numeric. Minimum number of tokens in flist_A.}

\item{min_B}{numeric. Minimum number of tokens in flilst_B.}

\item{freq_list_A}{data.table. Frequency list of tokens with columns "token", "N", "R"}

\item{freq_list_B}{data.table. Frequency list of tokens with columns "token", "N", "R"}
}
\value{
data.table with log likelihood and log ratio statistics.
}
\description{
\code{keywords} calculates keywords based on two frequency lists.
}
\examples{
\dontrun{keywords(my_flist_a, my_flist_b, A_is_subset = T)}

}
