% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_operations.R
\name{list_corpora}
\alias{list_corpora}
\title{List available corpora}
\usage{
list_corpora()
}
\description{
Lists corpora available through the Corpus Workbench. Make sure to set the
correct path to the corpus workbench registry directory before loading rcqp
or rusecqp (\code{Sys.setenv(CORPUS_REGISTRY = path_to_reg)}).
}
\examples{
\dontrun{list_corpora()}

}
