% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{q_collocations}
\alias{q_collocations}
\title{Collocations to query match}
\usage{
q_collocations(query_result, context, pattr = "word", exclude_match = T)
}
\arguments{
\item{query_result}{list of class query result returned from
\code{query_corpus}.}

\item{context}{integer vector of length one: number of tokens around the
query match. integer vector of length two: number of tokens to the left and
right. character: structural attribute (e.g. "s").}

\item{pattr}{character specifying positional attribute (e.g. "word").}

\item{exclude_match}{logical indicating whether the matched tokens themselves
should be excluded from the analysis.}
}
\value{
data.table of collocators and frequencies.
}
\description{
\code{q_collocations} performs collocation analysis of a query match. If
query match is longer than one token the context is build around the entire
length of the query match. E.g. if \code{context = 3}, three tokens before
the first token of the match and three tokens after the last token of the
match are taken into account. In case two or more numeric collocation windows
overlap, tokens are counted multiple times. However, if more than one match
fall within a structural region, only unique structural regions are
considered.
}
\examples{
\dontrun{q_collocations(my_query_result, context = 3)}
  \dontrun{q_collocations(my_query_result, context = c(2, 0))}
  \dontrun{q_collocations(my_query_result, context = "s", pattr = "lemma", exclude_match = F)}

}
