% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{q_distribution}
\alias{q_distribution}
\title{Frequency distribution of query results}
\usage{
q_distribution(query_result, sattr)
}
\arguments{
\item{query_result}{list of class query result returned from
\code{query_corpus}.}

\item{sattr}{character of structural attribute (e.g. "text_year")}
}
\value{
frequency distribution as data.table of frequencies.
}
\description{
\code{q_frequency_breakdown} returns number of query matches in each
category of the structural attribute provided.
}
\examples{
\dontrun{q_distribution(my_query_result, "text_year")}

}
