% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{q_frequency_breakdown}
\alias{q_frequency_breakdown}
\title{Frequency breakdown of query results}
\usage{
q_frequency_breakdown(query_result, pattr = "word")
}
\arguments{
\item{query_result}{list of class query result returned from
\code{query_corpus}.}

\item{pattr}{character of positional attribute (e.g. word)}
}
\value{
data.table of frequencies.
}
\description{
\code{q_frequency_breakdown} returns a frequency list of query results.
}
\examples{
\dontrun{q_frequency_breakdown(my_query_result, pattr = "lemma")}

}
