% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{query_corpus}
\alias{query_corpus}
\title{Query corpus}
\usage{
query_corpus(cqp_corpus, cqp_query, sattr = NULL, sattr_values = NULL)
}
\arguments{
\item{cqp_query}{character with cqp query string.}

\item{sattr}{character specifying s-attribute.}

\item{sattr_values}{vector of s-attribute values.}

\item{corpus}{corpus created with \code{get_corpus}.}
}
\value{
list of class query_result with match and matchend position values.
}
\description{
\code{query_corpus} performs a cqp_query and returns a list of position
values. If s_attr and s_attr_values are provided, restricts query to
structural attributes with given values.
}
\examples{
\dontrun{query_corpus(my_corpus, '[word = "crisis"]')}

}
