% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusecqp.R
\docType{package}
\name{rusecqp}
\alias{rusecqp}
\alias{rusecqp-package}
\title{rusecqp: A package for performing corpus linguistic analysis on Corpus
Workbench corpora}
\description{
A package of corpus analysis functions to analyze corpora indexed in the
Corpus Workbench. Uses rcqp cqi_ set of functions, but provides high abstract
functions to process query results, perform collocation and keyword analysis,
calculate ngrams, etc.
}
