% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_operations.R
\name{get_corpus}
\alias{get_corpus}
\title{Get information on existing Corpus Workbench corpus}
\usage{
get_corpus(corpus_name)
}
\arguments{
\item{corpus_name}{string of CWB corpus name.}
}
\value{
list with class attribute "rusecqp_corpus" containing  corpus information.
}
\description{
\code{get_corpus} retrieves corpus information from an existing CWB corpus.
\code{corpus_name} is converted to uppercase.
}
\examples{
\dontrun{get_corpus("MY_CORPUS")}
}
