% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{q_collocations}
\alias{q_collocations}
\title{Collocations to query match}
\usage{
q_collocations(query_result, context, pattr = "word", include_match = F)
}
\arguments{
\item{query_result}{list of class query result returned from
\code{query_corpus}.}

\item{context}{integer vector of length one: number of tokens around the
query match. integer vector of length two: number of tokens to the left and
right. character: structural attribute (e.g. "s").}

\item{pattr}{character specifying positional attribute (e.g. "word").}

\item{include_match}{logical indicating wether the matched tokens themselves
should be considered in the analysis.}
}
\value{
data.table of collocators and frequencies.
}
\description{
\code{q_collocations} returns collocators of a query match. If query match is
longer than one token, the context is build around the entire length of the
query match. E.g. if \code{context = 3}, three tokens before the first token
of the match and three tokens after the last token of the match are taken
into account.
}
\examples{
\dontrun{q_collocations(my_query_result, context = 3)}
  \dontrun{q_collocations(my_query_result, context = c(2, 0))}
  \dontrun{q_collocations(my_query_result, context = "s")}

}
